move internal used functions to internal
