+++ internal/controller/deployer/deployer_controller.go
@@ -27,9 +27,9 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	deliveryv1alpha1 "github.com/open-component-model/ocm-k8s-toolkit/api/v1alpha1"
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/ocm"
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/status"
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/util"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/ocm"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/status"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/util"
 )
 
 // Reconciler reconciles a Deployer object.
