+++ internal/controller/resource/resource_controller.go
@@ -49,9 +49,9 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	"github.com/open-component-model/ocm-k8s-toolkit/api/v1alpha1"
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/ocm"
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/status"
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/util"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/ocm"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/status"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/util"
 )
 
 type Reconciler struct {
