+++ internal/ocm/ocm_test.go
@@ -38,7 +38,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/open-component-model/ocm-k8s-toolkit/api/v1alpha1"
-	. "github.com/open-component-model/ocm-k8s-toolkit/pkg/ocm"
+	. "github.com/open-component-model/ocm-k8s-toolkit/internal/ocm"
 )
 
 const (
