+++ internal/status/register_status_defer.go
@@ -15,7 +15,7 @@ import (
 	eventv1 "github.com/fluxcd/pkg/apis/event/v1beta1"
 	kuberecorder "k8s.io/client-go/tools/record"
 
-	"github.com/open-component-model/ocm-k8s-toolkit/pkg/event"
+	"github.com/open-component-model/ocm-k8s-toolkit/internal/event"
 )
 
 // UpdateStatus takes an object which can identify itself and updates its status including ObservedGeneration.
