--- pkg/substitute/doc.go
@@ -1,2 +1,0 @@
-// This package contains our generic substitution engine that can be used to substitute values in files or data.
-package substitute
