--- pkg/substitute/steps/step.go
@@ -1,5 +1,0 @@
-package steps
-
-type Step interface {
-	Substitute(path string) error
-}
